move split
