+++ br/pkg/restore/import.go
@@ -22,6 +22,7 @@ import (
 	"github.com/pingcap/tidb/br/pkg/conn"
 	berrors "github.com/pingcap/tidb/br/pkg/errors"
 	"github.com/pingcap/tidb/br/pkg/logutil"
+	"github.com/pingcap/tidb/br/pkg/restore/split"
 	"github.com/pingcap/tidb/br/pkg/summary"
 	"github.com/pingcap/tidb/br/pkg/utils"
 	"github.com/pingcap/tidb/br/pkg/utils/utildb"
@@ -78,7 +79,7 @@ type ImporterClient interface {
 
 type importClient struct {
 	mu         sync.Mutex
-	metaClient SplitClient
+	metaClient split.SplitClient
 	clients    map[uint64]import_sstpb.ImportSSTClient
 	tlsConf    *tls.Config
 
@@ -86,7 +87,7 @@ type importClient struct {
 }
 
 // NewImportClient returns a new ImporterClient.
-func NewImportClient(metaClient SplitClient, tlsConf *tls.Config, keepaliveConf keepalive.ClientParameters) ImporterClient {
+func NewImportClient(metaClient split.SplitClient, tlsConf *tls.Config, keepaliveConf keepalive.ClientParameters) ImporterClient {
 	return &importClient{
 		metaClient:    metaClient,
 		clients:       make(map[uint64]import_sstpb.ImportSSTClient),
@@ -199,7 +200,7 @@ func (ic *importClient) SupportMultiIngest(ctx context.Context, stores []uint64)
 
 // FileImporter used to import a file to TiKV.
 type FileImporter struct {
-	metaClient   SplitClient
+	metaClient   split.SplitClient
 	importClient ImporterClient
 	backend      *backuppb.StorageBackend
 	rateLimit    uint64
@@ -212,7 +213,7 @@ type FileImporter struct {
 
 // NewFileImporter returns a new file importClient.
 func NewFileImporter(
-	metaClient SplitClient,
+	metaClient split.SplitClient,
 	importClient ImporterClient,
 	backend *backuppb.StorageBackend,
 	isRawKvMode bool,
@@ -299,8 +300,8 @@ func (importer *FileImporter) Import(
 		tctx, cancel := context.WithTimeout(ctx, importScanRegionTime)
 		defer cancel()
 		// Scan regions covered by the file range
-		regionInfos, errScanRegion := PaginateScanRegion(
-			tctx, importer.metaClient, startKey, endKey, ScanRegionPaginationLimit)
+		regionInfos, errScanRegion := split.PaginateScanRegion(
+			tctx, importer.metaClient, startKey, endKey, split.ScanRegionPaginationLimit)
 		if errScanRegion != nil {
 			return errors.Trace(errScanRegion)
 		}
@@ -379,9 +380,9 @@ func (importer *FileImporter) Import(
 				switch {
 				case errPb.NotLeader != nil:
 					// If error is `NotLeader`, update the region info and retry
-					var newInfo *RegionInfo
+					var newInfo *split.RegionInfo
 					if newLeader := errPb.GetNotLeader().GetLeader(); newLeader != nil {
-						newInfo = &RegionInfo{
+						newInfo = &split.RegionInfo{
 							Leader: newLeader,
 							Region: info.Region,
 						}
@@ -403,7 +404,7 @@ func (importer *FileImporter) Import(
 						logutil.Region(info.Region),
 						zap.Stringer("newLeader", newInfo.Leader))
 
-					if !checkRegionEpoch(newInfo, info) {
+					if !split.CheckRegionEpoch(newInfo, info) {
 						errIngest = errors.Trace(berrors.ErrKVEpochNotMatch)
 						break ingestRetry
 					}
@@ -454,7 +455,7 @@ func (importer *FileImporter) setDownloadSpeedLimit(ctx context.Context, storeID
 
 func (importer *FileImporter) downloadSST(
 	ctx context.Context,
-	regionInfo *RegionInfo,
+	regionInfo *split.RegionInfo,
 	file *backuppb.File,
 	rewriteRules *RewriteRules,
 	cipher *backuppb.CipherInfo,
@@ -526,7 +527,7 @@ func (importer *FileImporter) downloadSST(
 
 func (importer *FileImporter) downloadRawKVSST(
 	ctx context.Context,
-	regionInfo *RegionInfo,
+	regionInfo *split.RegionInfo,
 	file *backuppb.File,
 	cipher *backuppb.CipherInfo,
 ) (*import_sstpb.SSTMeta, error) {
@@ -593,7 +594,7 @@ func (importer *FileImporter) downloadRawKVSST(
 func (importer *FileImporter) ingestSSTs(
 	ctx context.Context,
 	sstMetas []*import_sstpb.SSTMeta,
-	regionInfo *RegionInfo,
+	regionInfo *split.RegionInfo,
 ) (*import_sstpb.IngestResponse, error) {
 	leader := regionInfo.Leader
 	if leader == nil {
