+++ br/pkg/restore/range.go
@@ -9,13 +9,13 @@ import (
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/kvproto/pkg/import_sstpb"
-	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/log"
+	"go.uber.org/zap"
+
 	berrors "github.com/pingcap/tidb/br/pkg/errors"
 	"github.com/pingcap/tidb/br/pkg/logutil"
 	"github.com/pingcap/tidb/br/pkg/rtree"
 	"github.com/pingcap/tidb/tablecodec"
-	"go.uber.org/zap"
 )
 
 // Range record start and end key for localStoreDir.DB
@@ -103,20 +103,6 @@ func SortRanges(ranges []rtree.Range, rewriteRules *RewriteRules) ([]rtree.Range
 	}
 	sortedRanges := rangeTree.GetSortedRanges()
 	return sortedRanges, nil
-}
-
-// RegionInfo includes a region and the leader of the region.
-type RegionInfo struct {
-	Region *metapb.Region
-	Leader *metapb.Peer
-}
-
-// ContainsInterior returns whether the region contains the given key, and also
-// that the key does not fall on the boundary (start key) of the region.
-func (region *RegionInfo) ContainsInterior(key []byte) bool {
-	return bytes.Compare(key, region.Region.GetStartKey()) > 0 &&
-		(len(region.Region.GetEndKey()) == 0 ||
-			bytes.Compare(key, region.Region.GetEndKey()) < 0)
 }
 
 // RewriteRules contains rules for rewriting keys of tables.
